+++ src/index.js
@@ -1,10 +1,13 @@
 import { hijackEffects } from 'stop-runaway-react-effects';
+import { getVisibleItems } from './utilities/visibleItems';
 import ApolloClient from './components/apolloClient';
 import Button from './components/button';
 import Field from './components/field';
+import Hamburger from './components/hamburger';
 import HelperText from './components/helperText';
 import Input from './components/input';
 import Label from './components/label';
+import Menu from './components/menu';
 import PrivateRoute from './components/privateRoute';
 import Setup from './components/setup';
 
@@ -13,12 +16,15 @@ if (process.env.NODE_ENV !== 'production') {
 }
 
 export {
+  getVisibleItems,
   ApolloClient,
   Button,
   Field,
+  Hamburger,
   HelperText,
   Input,
   Label,
+  Menu,
   PrivateRoute,
   Setup,
 };